split out decryptReader and packerManager
